+++ applications/pyhdi/hdi_wrapper.h
@@ -24,22 +24,22 @@ class HDIParameters {
 protected:
     HDILogger *_logger = nullptr;
 
-    double *_source = nullptr;
-    double *_target = nullptr;
+    float *_source = nullptr;
+    float *_target = nullptr;
     unsigned int _n_points = 0;
     unsigned int _source_dimensions = 0;
     unsigned int _target_dimensions = 0;
 
-    double _perplexity = 30.0;
+    float _perplexity = 30.0;
     int _seed = 0;
-    double _minimum_gain = 0.1;
-    double _eta = 200.;
-    double _momentum = 0.5;
-    double _final_momentum = 0.8;
-    double _mom_switching_iter = 250;
-    double _exaggeration_factor = 4.0;
+    float _minimum_gain = 0.1;
+    float _eta = 200.;
+    float _momentum = 0.5;
+    float _final_momentum = 0.8;
+    float _mom_switching_iter = 250;
+    float _exaggeration_factor = 4.0;
     unsigned int _remove_exaggeration_iter = 250;
-    double _theta = 0.5;
+    float _theta = 0.5;
 
 public:
 
@@ -52,27 +52,27 @@ public:
 
     void set_n_points(unsigned int value);
 
-    void set_input(double *matrix, unsigned int n_points, unsigned int dimensions);
+    void set_input(float *matrix, unsigned int n_points, unsigned int dimensions);
 
-    void set_output(double *matrix, unsigned int n_points, unsigned int dimensions);
+    void set_output(float *matrix, unsigned int n_points, unsigned int dimensions);
 
-    void set_perplexity(double value);
+    void set_perplexity(float value);
 
     void set_seed(int value);
 
-    void set_minimum_gain(double value);
+    void set_minimum_gain(float value);
 
-    void set_eta(double value);
+    void set_eta(float value);
 
-    void set_theta(double value);
+    void set_theta(float value);
 
-    void set_momentum(double value);
+    void set_momentum(float value);
 
-    void set_final_momentum(double value);
+    void set_final_momentum(float value);
 
-    void set_mom_switching_iter(double value);
+    void set_mom_switching_iter(float value);
 
-    void set_exaggeration_factor(double value);
+    void set_exaggeration_factor(float value);
 
     void set_remove_exaggeration_iter(unsigned int value);
